od: Bug fix in seismic data converter when survey is changed from OD.
Summary:
ObjectSet of 2D IOPars is not deleted as a local static object is created for doing the job. So deepErase( all2dseisiopars_ ) is moved from destructor to doConvesion().

Files should be added to the delete list when the corresponding entry in .2ds has been made.

Made more checks before making a change in .2ds file.

Test Plan: tested with 3 different surveys.

Reviewers: raman

Reviewed By: raman

Maniphest Tasks: T104

Differential Revision: https://phabricator.dgbes.com:8443/D3537

git-svn-id: https://svn.opendtect.org/od/trunk@34003 a985855d-d384-4b45-ae33-31a410998c3c