od4.6: Removed large 'xcf' files that are not used but are released.
Summary:
When cleaning up icons I found these files, and they are huge. They are like the 'source code' for some icons, but should not be distributed in the release.
Permission requested to aslo remove these from 'od'.

Test Plan: Did a search for 'xcf' in OD but nothing found.

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D3560

git-svn-id: https://svn.opendtect.org/od/branches/od4.6@34029 a985855d-d384-4b45-ae33-31a410998c3c