Implemented uiString support in the tasks.
Summary:
Nanne: I don't expect you to OK this one (and it is not checked enough). I submit it to start the discussion on what to do.

There is a major flaw in our current Task::message(). The problem is that it is not syncronized. That is, there is no guarantee the const char* returned is valid. One could address that by making it return a BufferString. As we are changing with the uiStrings, I thought it would be natural to do it that way.

All calls to message() and nrDoneText() are removed in this commit. The (potential) problem is that we end up with a mixture of uiString and const char* for error messages.

In this commit the SeisReader/Writer uses uiString for errMsg(). But there are many classes that looks at them, and they still use const char* (or bufferString).

Test Plan: Not tested. od compiles, links and starts up.

Reviewers: aneesh

Reviewed By: aneesh

CC: helene, nanne

Maniphest Tasks: T163

Differential Revision: https://phabricator.dgbes.com:8443/D3487

git-svn-id: https://svn.opendtect.org/od/trunk@34082 a985855d-d384-4b45-ae33-31a410998c3c