od: Error messages from Geometry2D::write() will go to uiMSG().
Summary:
If in writing 2D geometries something went wrong then the error messages will go to uiMSG() and not just in the console.

Also when converting OD4 geometries to OD5 geometries if conversion is already done, geometry will not be written again to the file. So will now write geometry only once.

Test Plan: tested with F3.

Reviewers: raman

Reviewed By: raman

Differential Revision: https://phabricator.dgbes.com:8443/D3594

git-svn-id: https://svn.opendtect.org/od/trunk@34087 a985855d-d384-4b45-ae33-31a410998c3c