od : Fixing ambiguity in case of operator[]
Summary:
CDash fails on windows 32 bit.

http://intranet/cdash/viewBuildError.php?buildid=34328
e:\users\dgb\dev\buildscripts\od\build\dgb109_win32_od_debug_nightly_3689\od\include\basic\arraynd.h(334) : error C2666: 'ArrPtrMan<T>::operator []' : 3 overloads have similar conversions


The problem here is that ArrPtrMan class has operator type*() and operator[] (with od_int64 as argument).
So as long as we use ptr[num] and num is of type od_int64, there is no ambiguity. But when num is of type say int the compiler has to make implicit conversion. In Such case it has choice to cast num to od_int64 and use operator[] or use type*().

The problem arises whenever it has to cast the argument. So to avoid this explicitly added overloaded operator[] with different type of arguments( size_t, int, od_int64 ).

The error is bit strange only in case of 32 bit windows. For more details see
http://stackoverflow.com/questions/1726740/c-error-operator-2-overloads-have-similar-conversions 

Is there a better way to fix this error ?

Test Plan: Compiled on Windows( 32 and 64 ) and on linux without error or warning.

Reviewers: kristofer

Reviewed By: kristofer

CC: raman

Differential Revision: https://phabricator.dgbes.com:8443/D3608

git-svn-id: https://svn.opendtect.org/od/trunk@34108 a985855d-d384-4b45-ae33-31a410998c3c