Enabled uiZAxisTransform classes based on wells.
Summary:
This is a part of T153 where we are unifying the selection tools for selection
of z-transforms. The Well functionality is re-implemented in a tool that
inherits uiZAxisTransform, and the uiZAxisDataTransform takes the place of
uiT2DConvSel in the HorizonExport dlg.

Test Plan: Tested exporting transformed horizons, which works well. Also tested making depth-scenes based on the wells, which works well.

Reviewers: satyaki

Reviewed By: satyaki

Maniphest Tasks: T2, T153

Differential Revision: https://phabricator.dgbes.com:8443/D3589

git-svn-id: https://svn.opendtect.org/od/trunk@34149 a985855d-d384-4b45-ae33-31a410998c3c