Make render hud view available.
Summary:
The changes are used to do offscreen renderer hud view to image. and save the image with main view image as one image.
There are a little bit painful for doing ref and unref on osg::Image. But it seems no simple way to avoid these if we want to use 
sub function to do offscreen renderer.

Test Plan: Compiles and Tests working fine

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D3610

git-svn-id: https://svn.opendtect.org/od/trunk@34172 a985855d-d384-4b45-ae33-31a410998c3c