od : Fixing crash at multiple places. Calling uiPickSetMgr::storeNewSet instead of uiPickSetMgr::storeSetAs.
Summary:
Earlier in 4.6 uiPickPartServer inherited uiPickSetMgr and storeNewSet was called at all these places.
In od uiPickSetMgr& uipsmgr_ is a member of uiPickSetPartServer. storeNewSet is 
a protected function of uiPickSetMgr, so it cannot be directly acessed using uipsmgr_.storeNewSet( newps ). 

In your last commit Revision 34142
( Changes to import PickSet. Made dialog non-modal and added immediate display.
Also cleaned up the pickpartserver a bit. )

storeSetAs was called insead of storeNewSet in all these places. But this leads to crash as storesetAs does not add things to the manager and it crashes while accessing it.

So used storeNewSet instead of storeSetAs, but for this I had to make uiPickPartServer a friend class of uiPickSetMgr ( see D3649 ).

Is this change OK ( including the changes related to this of D3649 ) ? Or should we make storeNewSet a public function or change things in StoreSetAs itself ?

Test Plan: Does not crash while creating polygon

Reviewers: nanne

Reviewed By: nanne

CC: raman

Differential Revision: https://phabricator.dgbes.com:8443/D3661

git-svn-id: https://svn.opendtect.org/od/trunk@34202 a985855d-d384-4b45-ae33-31a410998c3c