od: Include non-ui plugins to the list of products
Summary: The product for each plugin now should contain all the ui and non-ui plugins

Test Plan: Works, but there is one problem. Since the non-ui plugins are treated as early. Those plugins are already loaded when the uiPlugineSel dialog pops up. So even when that plugin is unchecked it comes back in the list of plugins selected, becuase it has already loaded before this step. The chnage takes place next time when the program is ran.

Reviewers: nanne

Reviewed By: nanne

CC: kristofer, raman

Maniphest Tasks: T166

Differential Revision: https://phabricator.dgbes.com:8443/D3715

git-svn-id: https://svn.opendtect.org/od/trunk@34318 a985855d-d384-4b45-ae33-31a410998c3c