Made HorIterator thread safer.
Summary:
Triggered by D3662, it was seen that the HorSamplingIterator was not threadsafe.This revision changes the internal funcitonality of the class, in that the classnow has a current position, while before, it took an incoming positions and moved that forward.

All places where this class is used has been inspected to ensure correct behavious.

Test Plan: All changes not tested, but visually inspected. New test is added to ensure class does what it should.

Reviewers: bert

Reviewed By: bert

CC: helene

Differential Revision: https://phabricator.dgbes.com:8443/D3716

git-svn-id: https://svn.opendtect.org/od/trunk@34326 a985855d-d384-4b45-ae33-31a410998c3c