Bugfix: 2D line seismics did no longer adapt to change of display properties.
Summary:
2D line seismics did no longer adapt to change of display properties, after 
revision 34129 which added polylines in top of survey box. Reverted changes
concerning the material of the panelstrips.

Test Plan:
2D line seismics adapts to change of display properties again. Do not see any
bad consequences in the scene after reverting this. Please try to remember
whether there was a specific reason for this part of your modification, in
which case I can test that and find a solution which covers it all.

Reviewers: nanne

Reviewed By: nanne

CC: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D3724

git-svn-id: https://svn.opendtect.org/od/trunk@34331 a985855d-d384-4b45-ae33-31a410998c3c