Made transformTrc mandatory.
Summary: I made transformTrc the normal way of operating the Z-transforms, using the new, unified way of getting the positions. The change is not fully tested yet, but I wish to move it in sooner than later.

Test Plan:
The following has to be tested:

# Batch transforms
# All tranforms involving reading data, especially in 2D.

A number of utility functions were added, both on the Survey::GeometryManager, as well as on the EM::Horizon.

Reviewers: raman

Reviewed By: raman

Maniphest Tasks: T189

Differential Revision: https://phabricator.dgbes.com:8443/D3743

git-svn-id: https://svn.opendtect.org/od/trunk@34383 a985855d-d384-4b45-ae33-31a410998c3c