Restored tracker planes and ortho slices based on osgGeo functionality
Summary:
Restoration of tracker planes and ortho slices. Makes use of the multi-slice
functionality of the osgGeo::LayeredTexture class (D3544). Main difference
compared to Coin release is that one ortho-slice with three selectable
orientations replaces three ortho-slices with a fixed orientation.

Test Plan:
Heavily tested in the context of the multi-slice LayeredTexture functionality.
More testing required in the context of horizon tracking workflows. Yet to be
committed because of upcoming API-freeze.

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D3758

git-svn-id: https://svn.opendtect.org/od/trunk@34409 a985855d-d384-4b45-ae33-31a410998c3c