Fix for compiler bug on Mac/XCode.
Summary:
Looks ugly, and I hope to remove it soon, but this will make it possible to work more efficciently.

There is a real danger to enable __MAC_LLVM_COMPILER_ERROR__ when this compiler is found, as we may then compile the release on an older release, and then a plugin-builder will build on a newer system. We may then have different ABI for theobjectset and the one that is built. Hence, it should only be enabled on purpose.

Test Plan: OD builds with this change, if I set the __MAC_LLVM_COMPILER_ERROR__ variable in CMakeCache.txt

Reviewers: bert

Reviewed By: bert

Subscribers: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D3785

git-svn-id: https://svn.opendtect.org/od/trunk@34441 a985855d-d384-4b45-ae33-31a410998c3c