od: Load all plugins after popup dialog / filter based on modulename
Summary:
As in title. As far as I can see there's no problem in loading plugins after the popup dialog.
It's much cleaner to fill the dontloadlist with the module name. Even better would be the product-name, but then the product needs a proper identifier.

Test Plan: OpendTect starts and the plugin deselection works better

Reviewers: ranojay

Reviewed By: ranojay

Subscribers: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D3719

git-svn-id: https://svn.opendtect.org/od/trunk@34472 a985855d-d384-4b45-ae33-31a410998c3c