od: Crash due to null array pointers.
Summary: Crash while examining autocorrelation output in GapDecon attribute window. Crash in cases where inputdata_->nrsamples_ is zero or when computation is for autocorrelation ( ngap_ and nlag_ are zero in this case as lagsize_ = 0 and gapsize_ = 0. )

Test Plan: This resolves some crashes, still evaluation of autocorrelation is not possible due to some other crash.

Reviewers: helene

Reviewed By: helene

Subscribers: kristofer, helene, raman

Differential Revision: https://phabricator.dgbes.com:8443/D3828

git-svn-id: https://svn.opendtect.org/od/trunk@34508 a985855d-d384-4b45-ae33-31a410998c3c