od: visEevent crash fixed, which was only appearing on Windows 32 bit.
Summary:
The callback function should always be declared with a CallBacker* argument. Because later when it is actually called form  the Callback class, it always calls the function with an argument e.g.

if ( obj_ && fn_ )
	(obj_->*fn_)( cber ); // always argument

So if the function is declared without this argument, and is called with the argument, on Widows it results in the violation calling convention.
Surprisingly no other platform or compiler reported this problem before.

Test Plan: Crash fixed

Reviewers: jaap

Reviewed By: jaap

Subscribers: kristofer, raman

Differential Revision: https://phabricator.dgbes.com:8443/D3894

git-svn-id: https://svn.opendtect.org/od/trunk@34600 a985855d-d384-4b45-ae33-31a410998c3c