od: Added option to set scale of inl/crl/z axis
Summary:
User friendliness request. Added option to set scale of inl/crl/z axis in 3D scene.
Made this available in the Scene properties window.
The scale is not saved in the session yet.

A save in settings is difficult, because each survey has different scaling. This scale should go in a survey dependent settings file, whereas the other parameters are saved in the general user settings.

Test Plan:
Opened scene properties window and changed the scale.
- Both scale text and the gridlines are updated correctly.
- On cancel the initial scale is set back.

Reviewers: jaap

Reviewed By: jaap

Subscribers: kristofer

Maniphest Tasks: T107

Differential Revision: https://phabricator.dgbes.com:8443/D3954

git-svn-id: https://svn.opendtect.org/od/trunk@34724 a985855d-d384-4b45-ae33-31a410998c3c