od5.0: Support sub-inputs in uiMathExpressionVariable and uiMathFormula
Summary:
To support multi-component seismic selection we needed some kind of sub-input mechanism.
That is now there, only happens when enabled in the Setup for uiMathFormula.
The input definitions need to be a FileMultiString of input`subinput.

Test Plan: The old stuff still works, this needs to be tested when Helene adds multi-component stuff in the Math Attribute part.

Reviewers: helene

Reviewed By: helene

Differential Revision: https://phabricator.dgbes.com:8443/D4002

git-svn-id: https://svn.opendtect.org/od/branches/od5.0@34780 a985855d-d384-4b45-ae33-31a410998c3c