od5.0: Fix for removal of 3D steering data
Summary: If there is a 2D attribute called 'Steering', a 2D dataset is created by that name. If a folder 'Steering' already exists under Seismics (for 3D steering cubes), the same is used to store the cbvs files. If you then removed the 'Steering' dataset from 'Manage seismics', the whole folder was removed (including the pre-existing 3D steering cubes).

Test Plan: Removed 'Steering' dataset, but the folder remains there if there are files in it.

Reviewers: nanne

Reviewed By: nanne

Maniphest Tasks: T104

Differential Revision: https://phabricator.dgbes.com:8443/D4011

git-svn-id: https://svn.opendtect.org/od/branches/od5.0@34786 a985855d-d384-4b45-ae33-31a410998c3c