od5.0: Support for other 'style' in FilePath parsing.
Summary:
The idea is to be able to parse 'any' style or rather 'mixed' style, irrespective of the platform. The immediate requirement of this was in reading the .2ds file created on another platform. So, if I do:

  FilePath fp( "/dsk/d2/surveys/F3_Demo_2012/Seismics" );
  fp.add( "Seis\\Seis^1.cbvs" );

The backslash on the second line should be parsed as a dir separator even on Linux.

Another place where this can be handy is in cross platform multi-machine processing. The paths can be stored in par files in any style and then they can be parsed by any platform.

Test Plan: Tested on linux by parsing a filepath with both '/' and '\'.

Reviewers: bert, nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D4009

git-svn-id: https://svn.opendtect.org/od/branches/od5.0@34787 a985855d-d384-4b45-ae33-31a410998c3c