od : Fix for Mantis Bug 2857 ( Unable to select a body made in Version 3.2 )
Summary:
see http://bugs.dgbes.com/view.php?id=2850
As described in the bug description 

"in F3 I added a body (Sandwave..) apparently generated in 3.2.
Asked if I wanted to convert ==Yes
Asked for a new name. Overwriting with same name kept asking: Select a body.
Changed the name (sandwave1): kept asking select body." 

User cannot select a body made in Version 3.2. The reason is that the translatorgroup name for the older versions of body was "MarchingCubesSurface" which has changed to "Body" in later version.

So in function uiIOObjSel::obtainIOObj() when we do this 
      const IOObj* ioob = iodir.get( inp.buf(),
				     workctio_.ctxt.trgroup->userName() );

IODir::get() function is called where we have 
	if ( ioobj->name() == ky )
	{
	    if ( !trgrpnm || ioobj->group() == trgrpnm )
		return ioobj;
	}

Here ioobj->group() is "MarchingCubesSurface" and trgrpnm is "body".
So because of this mismatch IODir::get always returns null, which is the cause of the error message.

The following changes fix this immediate problem by avoiding the trgrpnm which is an optional parameter. Is there a better fix for this ? Please suggest.

Test Plan: No more error message. Body is converted properly.

Reviewers: bert

Reviewed By: bert

Subscribers: kristofer, raman

Differential Revision: https://phabricator.dgbes.com:8443/D4024

git-svn-id: https://svn.opendtect.org/od/trunk@34830 a985855d-d384-4b45-ae33-31a410998c3c