od: Change of implementation of the instantaneous frequency attribute
Summary:
Replaced the current implementation by an implementation much closer to the actual formula. The output is now correctly constrained to the interval [0, Fnyquist]. This implementation will also be used in SeisTrcPropCalc::getFreq.
Further clean-ed the code, which was not always looking for undefs.

Test Plan: tested on three different surveys.

Reviewers: helene

Reviewed By: helene

Subscribers: kristofer, helene

Differential Revision: https://phabricator.dgbes.com:8443/D4074

git-svn-id: https://svn.opendtect.org/od/trunk@34862 a985855d-d384-4b45-ae33-31a410998c3c