od: More work on annotation arrow and image items.
Summary:
Finished remaining work for the annotation items
Please also see https://phabricator.dgbes.com:8443/D3879, where the initial code was copied from the annotation plugin to od main branch.

Note:
It seems the pointer returned by the qimage_->bits(), is used/manipulated by the osg::Image. So I have tried to make a local copy of this buffer and then pass it to osg::Image. I am not sure about this method and have doubts about the stability. So please let me know if this was a good way to do it or may there should be some other ways.

Test Plan: The items work as expected.

Reviewers: kristofer

Reviewed By: kristofer

Subscribers: kristofer, raman, nanne

Maniphest Tasks: T101

Differential Revision: https://phabricator.dgbes.com:8443/D4082

git-svn-id: https://svn.opendtect.org/od/trunk@34897 a985855d-d384-4b45-ae33-31a410998c3c