od : Fix for crash when StringListInpSpec::text function is called ( called when calculating Attributes )
Summary:
getFullString() function of uiString looks like...
BufferString		getFullString() const;

So it returns a BufferString which is immediately deleted after this function call. So returning strings_[cur_].getFullString().buf() would return a pointer of temporary object that would immediately be destroyed.
Accessing this deleted memory( const char* ) leads to crash while attribute calculations.   
  
Hence used static string mDeclStaticString( ret ) to return a valid const char*.The one down side is that multiple calls to this function would change the string leaving the initial caller with may be totally different string.

As 'strings_' is a member of this class so an alternative would be to simply return strings_[cur_].getOriginalString() that retuens a const char*.

Test Plan: No more crash

Reviewers: helene, kristofer

Reviewed By: kristofer

Subscribers: kristofer, raman

Differential Revision: https://phabricator.dgbes.com:8443/D4123

git-svn-id: https://svn.opendtect.org/od/trunk@34913 a985855d-d384-4b45-ae33-31a410998c3c