od5.0: Using const OD::String& when appropriate (replaces D4168)
Summary:
We are trying to avoid 'silent' upgrade of non-objects to BufferStrings, by using references to the OD::String class which is pure virtual.
A downside is that if a programmer makes a const BufferString& variable from such a const OD::String& then the compiler will silently create a temporary BufferString.

Test Plan: At first testing did get a crash because of the silent temporary probem, but I think I've hunted these down. Tested quite some work flows. Not sure how to avoid this problem creeping in in the future.

Reviewers: kristofer

Reviewed By: kristofer

Subscribers: helene, aneesh

Differential Revision: https://phabricator.dgbes.com:8443/D4183

git-svn-id: https://svn.opendtect.org/od/branches/od5.0@35006 a985855d-d384-4b45-ae33-31a410998c3c