od : Bug Fix for Mantis Issue 2905 ( Sampling data is not set when selecting a velocity cube )
Summary:
In a previous rework 'rangefld_' was moved to the base class.
uiVelModelZAxisTransform inherits uiTime2DepthZTransformBase which already has 'rangefld' as a member 
   uiZRangeInput*	rangefld_;

In uiVelModelZAxisTransform::setZRangeCB( CallBacker* ) there is a condition
    if ( !rangefld_ )
	return;
As rangefld_ was initialized with zero in the constructor of uiVelModelZAxisTransform, so here rangefld_ is always zero and function always returns without setting the range. Hence removed the member 'rangefld_' from uiVelModelZAxisTransform class.

Test Plan: The sampling data is set correctly ( compared with 4.6 version ).

Reviewers: kristofer, ding

Reviewed By: ding

Subscribers: kristofer, raman

Differential Revision: https://phabricator.dgbes.com:8443/D4203

git-svn-id: https://svn.opendtect.org/od/trunk@35015 a985855d-d384-4b45-ae33-31a410998c3c