od4.6 : Fix for Mantis Issue 2902 : [Crash Report] uiAttribCrossPlot::getSelectedID()
Summary:
see http://bugs.dgbes.com/view.php?id=2902
The crash happens when the list is empty and in that case 
attrsfld_->currentItem() returns -1. 
     const int curitem = attrsfld_->currentItem(); // curitem = -1;
This case was not handled in the code.

So const bool isstored = curitem >= attrsz; 
'isstored' here is false as 'curitem' is -1 and 'attrsz' is zero.
So the control reaches the else branch where attrinfo_->attrids_[curitem] was accessed. Here attrinfo->attrids_ size was zero and we were accessing its -1 element.
 
So added checks to avoid accessing invalid index and avoid crash.

Test Plan:
1. Tested the case when the listbox has no elements. No crash in this case.
2. Also tested the case when listbox has some items( to make sure there are no side effects of this change ).

Reviewers: helene

Reviewed By: helene

Subscribers: helene, raman

Differential Revision: https://phabricator.dgbes.com:8443/D4224

git-svn-id: https://svn.opendtect.org/od/branches/od4.6@35030 a985855d-d384-4b45-ae33-31a410998c3c