od : Bug Fix for Mantis Issue 2894 ( Copy Survey is constantly giving a warning message of invalid Directory )
Summary:
see http://bugs.dgbes.com/view.php?id=2894
The earlier design assumed that user can copy a survey from a Survey Dataroot. So if we want to copy a survey from a directory that is not a dataroot, the error message pops up.
Removed this constraint that the directory should be a dataroot to copy a survey.

Test Plan: Successfully copied a survey.

Reviewers: ranojay, raman

Reviewed By: raman

Subscribers: kristofer, raman

Differential Revision: https://phabricator.dgbes.com:8443/D4215

git-svn-id: https://svn.opendtect.org/od/trunk@35041 a985855d-d384-4b45-ae33-31a410998c3c