od5.0: Storing selections for 2D line selections (Fiexes Mantis 2887)
Summary:
Now also 2D line selections can be stored/retrieved.
In the process, made an object for the uiListBox filter.
Apllied that object to uiSelectFromList
Created a new object to pick 2D lines from a list
Used that object rather than uiSelectFromList in seis part server
And that fixes the problem with using the filter in the selection.

Test Plan: Tested both in od and od5.0

Reviewers: nanne

Reviewed By: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D4286

git-svn-id: https://svn.opendtect.org/od/branches/od5.0@35115 a985855d-d384-4b45-ae33-31a410998c3c