od : Fix for Mantis Issue 2920 ( Crash while Multi Attribute Selection )
Summary:
see http://bugs.dgbes.com/view.php?id=2920
As seen in the call stack in the constructor of uiMultiAttribSel there is a call to fillAttribFld() which  calls uiMultiAttribSel::entrySel.
And in uiMultiAttribSel::entrySel we have
     allcompfld_->setSensitive( outnames.size()>1 );
The reason for crash is at this time allcompfld_ is not initialized (It is initialized later in the constructor). Hence moving the call of fillAttribFld() later in the constructor after all UI objects have been created.

Test Plan: No more crash

Reviewers: helene

Reviewed By: helene

Subscribers: kristofer, helene, raman

Differential Revision: https://phabricator.dgbes.com:8443/D4351

git-svn-id: https://svn.opendtect.org/od/trunk@35164 a985855d-d384-4b45-ae33-31a410998c3c