Bug fixing 2918
Summary:
After investigation, when horizon is very small (less than cMinMemValSetParallelSize), the channel unmapped data is initialized by OD::memValueSet in which the size 
is done one more than original pointer's size. and then it will be crashed when delete the pointer in the destruction function where it is used.

Bert, Could you please help me to check it?

Test Plan: Compiles and Tests fine

Reviewers: bert

Reviewed By: bert

Differential Revision: https://phabricator.dgbes.com:8443/D4379

git-svn-id: https://svn.opendtect.org/od/branches/od5.0@35177 a985855d-d384-4b45-ae33-31a410998c3c