od : Bug Fix for Mantis Issue 2937 ( Horizon base color do not correspond to database )
Summary:
see http://bugs.dgbes.com/view.php?id=2937
As discussed the problem arises because of inconsistency in writing color to IOPar in the .2dh file and .par file.
In .2dh file color is written as an int ( rgb ) and in .par file it is written as a multistring. So to avoid this inconsisitency using IOPar's IOPar::set( const char* keyw, const Color& c ) and IOPar::get( const char* keyw, Color& c ) const. Also to read old file correctly where color is written as int made changes in IOPar::get to handle this case.

Test Plan: The horizon color is updated properly after the changes.

Reviewers: raman

Reviewed By: raman

Subscribers: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D4468

git-svn-id: https://svn.opendtect.org/od/trunk@35288 a985855d-d384-4b45-ae33-31a410998c3c