od5.0: Possible fix for Horizon intersection with random line not displayed in flattened scene.
Summary:
The Problem:
In a flattened scene the horizons laying on top of the flattend horizon does not seem to show the intersection 3d line at the point of intersection with the random track displays only.
http://bugs.dgbes.com/view.php?id=2854

What goes wrong:
The z-range for the random track displays were actually computed from the trace ranges and not from the CubeSampling of the randomtrack display. Thus it did not contain the actual Z-range, which in this case could also have negative values for a flattened scene. So a valid comparison for the values were failing and the display objects were not having valid coordinates.

Solution.
Taking the Z-range from the actual cube-sampling of the random-track display, where we now get the valid range, and the comparizon returns posetive results, so the display objects have valid coordinates.

Test Plan: Works as expected

Reviewers: jaap

Reviewed By: jaap

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com:8443/D4487

git-svn-id: https://svn.opendtect.org/od/branches/od5.0@35301 a985855d-d384-4b45-ae33-31a410998c3c