Modified /debug flag to be turned on only on releases
Summary: On windows, when compiling non-debug, pdb files are not generated. Therefor, as we need them in the release, we turned the /debug on. This does however slow down the continious integration, and therefor I am thinking of this change, where it will only be used during release-builds.

Test Plan: Not tested, but no regression should go unnoticed.

Reviewers: nageswara

Reviewed By: nageswara

Differential Revision: https://phabricator.dgbes.com:8443/D4495

git-svn-id: https://svn.opendtect.org/od/branches/od5.0@35315 a985855d-d384-4b45-ae33-31a410998c3c