od: Removed duplicate member variable
Summary: The same variable of the same type exists in (and is managed by) the base class (FlatDataPack) as well. And our compilers on all the platforms happily accepted it without even a warning. This made me write a test program (outside OD) and I discovered that C++ allows this across all platforms. What a pity!

Test Plan: The code still compiles and I see no reason why it should change anything functionally.

Reviewers: helene

Reviewed By: helene

Subscribers: kristofer, helene

Differential Revision: https://phabricator.dgbes.com:8443/D4516

git-svn-id: https://svn.opendtect.org/od/trunk@35335 a985855d-d384-4b45-ae33-31a410998c3c