Bug fixing 2940 about new wheeler scene displayed in wrong scale
Summary:
As discussed with Kristofer, we will use notifier to instead of scene timer in order to do viewAll for a new scene.
1. Added one notifier in vissurscene. it will trigger in the function addObject() that it is at least called once when creating a new scene. 
Sometimes it will be called more times if doing addObject() after creating a new scene. 
2. in uiodscenegmr, We subscriber the notifier of vissurscene and added the callback function in which it will do viewAll().
3. we changed the sceneTimer to tileTimer, because we need wait a while for finishing viewAll() render, and then do tile(), the render is relative in certain time, 
because the render always for a empty scene without od data.

Test Plan: Compiles and Tests fine

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D4506

git-svn-id: https://svn.opendtect.org/od/branches/od5.0@35356 a985855d-d384-4b45-ae33-31a410998c3c