od: Possible fix for horizon tracker not reloaded while retracking the horizon.
Summary:
The problem:
The horizon tracker always prompted the error message that the tracked horizon does not have a setup saved, though it already has a saved setup.
http://bugs.dgbes.com/view.php?id=2710

What goes wrong:
When the retrack procedure is initiatied, loadTrackSetupCB is called which calls   emtracker->getSectionTracker( sid, false ); the false argument specified does not create any tracker, so the list of trackers is empty. Later when the control reaches the useSavedSetupDlg, function there again it calls tracker->getSectionTracker( sid, true ), this time with true as argument, which creates the tracker here, but obviously it is not initialised, thus was unabale to read the setup file.

Solution:
I chnage the order of creation of the setup. Since the loadTrackSetupCB is called while creating the tracker, it would also create the section tracker during this call, so this time instead of calling the  emtracker->getSectionTracker( sid, false ); with false I do  emtracker->getSectionTracker( sid, true );. later in the useSavedSetupDlg(), here the stup should already be created, and I do not want to create it here, just check if it is there or has it been properly initialied. So I call emtracker->getSectionTracker( sid, false );

Test Plan: Works as expected

Reviewers: jaap, nanne

Reviewed By: nanne

Subscribers: kristofer, raman

Differential Revision: https://phabricator.dgbes.com:8443/D4503

git-svn-id: https://svn.opendtect.org/od/trunk@35376 a985855d-d384-4b45-ae33-31a410998c3c