Made uiString::arg() a template based function.
Summary:
There was a problem in the example of uiString:

uiString string = uiString( "%1 plus %2 is %3")
              .arg( toString(4) )
              .arg( toString(5) )
              .arg( toString(4+5) );

Where it turns out that the three toString is called in sequence, before calling arg. Hence theo utcome is wrong as .arg refers to a static buffer.

The solution is either to let toString return a BufferString, but that is also not very feasible.

Test Plan: Test is added to cover new scenario. Compiles & Links on Linux. Should be tested on Windows if approved, before committing.

Reviewers: nanne

Reviewed By: nanne

Subscribers: jaap, kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D4583

git-svn-id: https://svn.opendtect.org/od/trunk@35386 a985855d-d384-4b45-ae33-31a410998c3c