od4.6: Possible fix for crash while importing a horizon.
Summary:
The problem:
Crash occurs if the same horizon is imported which is already displayed in the scene.
http://bugs.dgbes.com/view.php?id=2599


What goes wrong:
Series of callbacks are triggered from the emobject while calling the setKnot function, which leads to the crash in Coin functions for refreshing the display.


The Solution:
Using a 2d array, which is populated with the Z-values, and ultimately setting this array to the horizon, instaed of calling setKnot for every position added.

Test Plan: crash fixed

Reviewers: raman

Reviewed By: raman

Differential Revision: https://phabricator.dgbes.com:8443/D4594

git-svn-id: https://svn.opendtect.org/od/branches/od4.6@35407 a985855d-d384-4b45-ae33-31a410998c3c