od4.6: Solved problems with auto-snapping to the wavelet center during import
Summary:
The reported index in the user message was wrong, and the center sample as set in WaveletAscIO::get. Luckily that was transparent because the import would propose to resnap. Both originate from large changes in the class made between od4.4 and od4.6.
Fixed both, by ensuring the user provides the row number in the file, as seen from the examine window, as the position of the center sample.
Also changed the auto-snapping in case of a quadrature wavelet: one should snap to the zero amplitude, not min/max
Also added a report of the average phase in the wavelet manager

Test Plan: tested in F3_Demo and on a wavelet from a support case

Reviewers: helene

Reviewed By: helene

Differential Revision: https://phabricator.dgbes.com:8443/D4656

git-svn-id: https://svn.opendtect.org/od/branches/od4.6@35426 a985855d-d384-4b45-ae33-31a410998c3c