od5.0: Preventing selection of arbitrary log for xplot
Summary:
When creating x-plot data I noticed that one log was selected, but it was different per survey. No idea which one, and why. Turned out that there is a default propperty name that is empty for simple x-plots. The 'nearest match' would then pick something.
At the same time, these list boxes are places where the user has to be forced to make a selection; the 'AtLeastOne' rule is not intuitive here. Also chaged that for xplot input.

Test Plan: tried in od and od5.0

Reviewers: satyaki

Reviewed By: satyaki

Subscribers: helene

Differential Revision: https://phabricator.dgbes.com:8443/D4675

git-svn-id: https://svn.opendtect.org/od/branches/od5.0@35458 a985855d-d384-4b45-ae33-31a410998c3c