od: Hide Jobname field
Summary:
Most users don't understand what the jobname field is. The ones who do understand, don't want to see it.
In the attribute output window, I hide the field, and set the jobname to output name.

I made a similar change to the SteeringCube window.
Should we apply it everywhere? And in that case, remove the jobfield completely and just work with a bufferstring to hold the jobname?

Test Plan: Job name is set correctly and processing still works fine.

Reviewers: raman, bert

Reviewed By: bert

Subscribers: kristofer, helene

Differential Revision: https://phabricator.dgbes.com:8443/D4485

git-svn-id: https://svn.opendtect.org/od/trunk@35459 a985855d-d384-4b45-ae33-31a410998c3c