od4.6: Added od_SeisMMBatch_console on Windows for debugging purposes.
Summary:
As discussed, I have added the od_SeisMMbatch_console. I was wondering if it would be a good idea to cotrol the on/off of this program with the environment variable DTECT_DEBUG, instead of DTECT_MM_DEBUG, which controls the verbosity of the od_SeisMMBatch.

Let me know your thoughts

Test Plan: Launched the multi-machine window, by setting and un-setting the environment variable DTECT_DEBUG, and the console and the non-console winodws appeared as expected.

Reviewers: nanne

Reviewed By: nanne

Subscribers: nageswara, raman

Differential Revision: https://phabricator.dgbes.com:8443/D4709

git-svn-id: https://svn.opendtect.org/od/branches/od4.6@35489 a985855d-d384-4b45-ae33-31a410998c3c