Bugfix: Adding or changing e.g. a horizon or fault attribute gave a default seismics color map instead of the data-cube specific default color map.
Summary:
(1) Adding or changing a random pos attribute (e.g. horizon or fault) gives
a default seismics color map instead of the data-cube specific default color
map. In the case of cube or trace attributes, this is covered by the obscure
"if (cacheid==DataPack::cNoID()) useDefColTab(visid,attrib)" command in
uiODApplMgr::getNewData(.), whose success very much depends on how data refresh
protocols are implemented in visSurvey objects. It does not work for horizons
and faults, mainly because there is no cache to be emptied.
A better approach seems to force the use of data-cube specific default color
maps at those places in the code where a new attribute is actually entering the
data tree. Because it is not sure whether this revision covers all these places
and because it won't hurt either, the old approach is prolonged as well.
(2) After solving (1), it turned out that in case of adding a new random pos
attribute, the color column in data tree was not updated until deselecting the
current attribute.

Test Plan:
Add a new fault or horizon to the scene and next add or change to e.g. the
ChimneyCube attribute. Now it comes with the default Chimney color map in
the scene and also in the color column of the data tree.

Reviewers: nanne

Reviewed By: nanne

Subscribers: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D4734

git-svn-id: https://svn.opendtect.org/od/trunk@35493 a985855d-d384-4b45-ae33-31a410998c3c