od : Bug Fix for Mantis Issue 2982 ( HC Crossplot - After reading the HC it gets stuck )
Summary:
http://bugs.dgbes.com/view.php?id=2982
When provider_->getData() returns null, then nrdone_ was not incremented.
So even when there is some processing going in the background user gets a feeling that there is no progress as progress bar is not updated.

Test Plan: The progress bar was updated properly when HC has smaller range than the specified sampling.

Reviewers: helene

Reviewed By: helene

Subscribers: kristofer, helene, raman

Differential Revision: https://phabricator.dgbes.com:8443/D4766

git-svn-id: https://svn.opendtect.org/od/trunk@35526 a985855d-d384-4b45-ae33-31a410998c3c