od5.0: uiIOSel: hook for adding external button
Summary:
In Petrel there is a 'blue arrow' button to fetch objects to a selection field. Whether we do it in exactly the same way is not relevant, but the key is that the button should be between the label and the input field.
This extension even allows more than one of these external gets.

Test Plan: still works without, but also tested by hacking a file to add a button there, which also worked.

Reviewers: bert, raman

Reviewed By: raman

Maniphest Tasks: T198

Differential Revision: https://phabricator.dgbes.com:8443/D4795

git-svn-id: https://svn.opendtect.org/od/branches/od5.0@35572 a985855d-d384-4b45-ae33-31a410998c3c